refactor: rename folder to maintain language consistency
